Add files via upload
The determination of optimal production planning is a combinatorial optimization problem and requires specific strategies for evolutionary computation techniques to be successful. The objective of the model is to identify an optimal production plan with maximum profit by satisfying all the associated constraints. In this work, we provide such a strategy that enables the use of evolutionary algorithms to efficiently solve the single level production planning. The efficacy of the strategy is demonstrated with the recently proposed Pelican Optimization Algorithm which also enables in evaluating the performance of this technique. It was observed that Pelican Optimization Algorithm is able to consistently solve the production planning problem with the suggested strategy. In this paper, we use a new stochastic nature-inspired optimization algorithm called Pelican Optimization Algorithm (POA). The main idea in designing the proposed POA is simulation of the natural behaviour of pelicans during hunting. In POA, search agents are pelicans that search for food sources. The mathematical model of capacity are determined using a piece–wise linear function between the two successive levels.